Use handle_validation_failure in postgres create route
Also, treat unrecognized flavor as standard flavor, instead of
if trying to handle it as an error.  This is only on the create
view, actual submissions will raise an error for invalid flavor.
You can only get the invalid flavor by manually changing the URL,
so there doesn't seem to be a need to make it an error, and
treating it as an error complicates the logic.