Require handle_validation_failure usage to pass specs
If a web spec runs and hits a code path that raises an error
without handle_validation_failure being called, the spec will now
raise an error. This alerts the developer to the problem, so they
can fix it by adding the appropriate handle_validation_failure
call.

In non-test mode, emit a log for these errors. We can use the
logged information to find routes where we don't have covering
specs for the failure case. Then we can add covering specs for
the failure case, and then add the appropriate
handle_validation_failure call to fix them.