Simplify base_url and request_path usage
These two instance variables are quite similar to each other. We don't need to
have both of them. We loose the ability to handle base_urls with parameters in
them, but we don't use the Mezmo endpoints with parameters anyway. We only use
`hostname`, which is already added by our client anyway, so caller doesn't need
to specify any query parameters in the base_url.