Use request.POST as fallback in cases where flash["old"] is not set
Apparently we did not have any tests for any of these cases. Add a
test that should exercise the radio_small_cards, resource_creation_form,
and textarea changes.

Do not set first_of_its_kind to true for checkbox inputs if the request
is not a GET request. This fixes a bug where an invalid SSH key format
(or other validation error) would recheck the enable IPv4 checkbox on
the VM creation form even if the user deliberately unchecked it.

Avoid redundant Vm.first calls in the spec while here.