Use correct params typecasting when handling validation failures
Previously, this would assume a given type, and likely raise
NoMethodError if the type was not what was expected. If there
is a typecast error when rendering the validation failure
template, display a simple error page.  This should only happen
if a user monkeys around with a form parameter before submitting
the request.

This uses a new feature in the Roda typecast_params plugin, so
update to the latest Roda commit.