Skip setting deadline for busy runners without workflow info
Once in a while, we don't receive the information on the workflows being
in_progress from Github via the webhooks. In these cases, before deciding
on canceling the runner, we check whether the runner is actually busy.
In this specific case, we should not set a deadline for archival as we
know that the runner is busy and could be running.