Use handle_validation_failure in the firewall attach/detach subnet route
This avoids the separate exception handling for web requests. The
api request handling is better here as it shows an error for the
specific field. It's also simpler. You can only hit an error in
the web case if the attached subnets are modified after the page
was displayed, or the user is manually changing them.