Remove VmHost#hetznerify
This was originally added in 1c75bfdb8e728127b97fde2ee49dec76258b1a65
(June 2023), and when originally added, it used HetznerHost, and it
worked correctly.

In b48756e163115bd36d5bb587c5314e7a48d4c772, it was changed from
using HetznerHost to using HostProvider, and at this point it broke,
because host_provider.{server_identifier,provider_name} are primary
key columns, and cannot be set via mass assignment by default.

It's possible to fix this issue, but considering it has been broken
for over 6 months, and is not used anywhere, it seems better to
remove it. The only reason this wasn't discovered sooner is that the
VmHost model specs are mostly mocked (yet another bug hidden by
mocking).