Replace system with sh in most cases in Rakefile
I found that monitor_smoke_test has been failing since the change
to log the first 5 up calls. However, it wasn't caught, because the
Rakefile used system, which only returns true/false by default.

We could use exception: true, but it's probably better to be
consistent with the rest of the Rakefile and use sh, which also
echoes the command run.