Do not allow removing current partition in the repartitioner
Fixes the following nondeterministic spec failure:

PG::InvalidTextRepresentation: ERROR:
invalid input syntax for type uuid: "100000000-0000-0000-0000-000000000000"

This could previously happen in the respirate specs due to a race
condition when the listen timeout was set to 0.01. Obviously, if
the current process is running, it should be considered active,
even if we haven't received a NOTIFY about it in the expected
timeframe.