Change invoice serializer return type from hash to struct
This makes it so indexing into the return value with an invalid
key results in an exception being raised instead of nil. This
additional strictness found a bug in the existing implementation,
where aggregated results would pass nil as the resource type and
resource family to BillingRate.line_item_usage, resulting in
incorrect display for InferenceTokens resource type, and
potentially suboptimal display (no conversion to integer) for
GitHubRunnerMinutes.

By switching to struct, this also allows switching from using
the [] method to calling normal methods.