Switch from using [] to using member methods for when using the invoice serializer
This results in code that is easier to read, and better indicates
the code is dealing with an object and not a hash.

Rename ser to data in some Invoice methods to keep it consistent.

Allow Invoice#generate_pdf argument to be optional, and use the
invoice serializer if the argument is not given.

Remove unnecessary detailed: true options when calling the invoice
serializer.

Avoid creating an unnecessary invoice serializer in the invoice
route.