Suppress invalid IAM instance profile name error while creating AWS instance
While provisioning a virtual machine at AWS, even if an IAM instance
profile is created, the create_instances command may fail due to the
following exception:

    Aws::EC2::Errors::InvalidParameterValue - Value
    (gr8x4mbvjdey2mdxh1fn7zzgtr-instance-profile) for parameter
    iamInstanceProfile.name is invalid. Invalid IAM Instance Profile
    name

This error is intermittent and disappears when the instance profile is
successfully propagated. We can ignore this error and retry the instance
creation, which makes the logs cleaner.