Add pg modify to the cli/sdk
This allows resizing or changing the HA type or tags for a postgres
resource.

This is the first cli command to deal with postgres tags. Postgres
tags are internally stored as a jsonb array containing hashes that
include "key" and "value" keys. There isn't an obvious approach for
accepting this on the command line. I chose the following format:

```
key1=value1,key2=value2
```

This should work for cases where keys don't contain `=` and
neither keys nor values contain `,`.