Add pg modify-firewall-rule to the cli/sdk
This allows you to modify an existing firewall rule. You can modify the
cidr and/or description for the rule.

While implementing this, I noticed the api endpoint always required a cidr
parameter, which made it more difficult to only modify the description.
Additionally, if you provided a cidr without a description parameter, it
would erase the current description, which seems unlikely to be the
desired behavior. This makes both the cidr and description parameters
optional, and if the parameter is not given, treat it the same as
providing the current parameter value.