Add pg add-config-entries to the cli/sdk
This also changes the api to make both pg_config and
pg_bouncer_config optional parameters instead of required
parameters for PATCH requests. The code already assumed the
parameters were optional.

Update the route code to require both pg_config and
pgbouncer_config be set for POST requests. This isn't
strictly required as committee enforces it, but in general
we treat committee as an extra layer of protection and not
as the sole validator.