Create KubernetesNode objects in provision_kubernetes_node
With this commit we will create a corresponding KubernetesNode for
each Vm but the progs still rely on the Vm objects itself.

This commit ensures we have the KubernetesNode objects for the newly
created cluster Vms. With another migration script, present Vms will
have their own KubernetesNode objects and strands.

Another commit will change the logic to rely on the KubernetesNode
object instead of Vms.