Use more accurate error for object reference containing ubid
Previously, it would treat an argument containing a ubid as a ubid,
even if it had more than 26 characters.

Add a test that checks that a string longer than 26 characters that
contains a string in ubid format is not treated as a ubid.

Without this change, you would get a "no * with id * exists" error.
With this change, you get a "invalid * reference (*), should be in
location/*-name or *-id format" error.