Combine the firewall_rule api-specific routing file into the main firewall route file
For some reason, we didn't do this when merging the routing trees
originally.  Most of the logic is the same, so it makes sense to
combine these routing files.

One thing I found is that we weren't consistent in the audit
logging. This attempts to make audit logging consistent, always
logging both the firewall and the firewall rule.