Fix validation/generation account names for social login
The social login may already have a name that it is invalid, so don't
just munge the email, munge the name instead if it is present.

Remove excess whitespace in the name. Limit name to 63 characters. If
the name would be empty, use Unknown as the name. Do not capitalize
the name, as that has the effect of removing capitalization from
characters other than the first.