Upgrade redocly cli to v2
Recently, Redocly CLI v2 was released with some breaking changes.
Dependabot opened a PR to upgrade it, but the linter failed on our API
spec with the following error [^1]:

    [1] openapi/openapi.yml:2585:21 at #/components/responses/LoadBalancer/content/application~1json/schema/allOf/1/required/0

    Required property 'location' is undefined.

    2583 | additionalProperties: false
    2584 | required:
    2585 |   - location
    2586 |   - subnet
    2587 |   - vms

While reviewing the migration guide from v1 to v2 [^2], I found that v2
introduces a new `spec` ruleset that enforces OpenAPI specification
compliance. Adding this ruleset to our configuration resolved the issue
and allowed the linter to pass.

[^1]: https://github.com/ubicloud/ubicloud/actions/runs/17173012435/job/48725151615?pr=3754
[^2]: https://redocly.com/docs/cli/guides/migrate-to-v2#spec-ruleset