Attempt to fix repartitioner spec nondeterminism
This attempts to fix the following nondeterministic spec failure:

```
  1) Repartitioner#notify uses NOTIFY to notify listeners on monitor channel
     Failure/Error: expect(th.value).to eq "1"

       expected: "1"
            got: "2"
```

One possible cause is there is still a thread notifying on the
channel before the spec starts. I looked at the other specs and
could not see where that was happening, though. However, one way
to work around this is to have the spec listen and notify on
a different channel.