Add bin/check-needed-openapi-paths
This runs the tests separately for each path in the openapi.yml file.
It removes the path entry in openapi.yaml, and runs the api specs.
If the api specs do not fail, that means the path isn't used and
shouldn't be in the openapi.yml file.

After running this, the openapi.yml file should only have paths
that are used by the application.