Print list of leaked threads after every spec
In general, specs should clean up (join) the threads they create.
If they don't, that's a bug to be fixed, because it can result
in other specs failing due to unexpected concurrency.

This currently prints a warning for leaked threads, but it may be
a good idea to switch it to an error at some point.