Switch firewall to use object/show view
Currently, this uses 3 tabs, overview, networking, and settings.

This change exposed that the display with limited permissions was
undertested, so this adds appropriate tests.

We previously were not checking that the currently attached
private subnets are actually viewable by the logged in user before
linking to them. This could result in a link that takes you in an
unauthorized access page. Check whether the private subnet is
viewable by the logged in user, and only link to it if it is viewable.