Enforce password length requirements
Previously, without the super call, the min/max length requirements
were ignored.

The related spec was not testing the password was actually changed,
since the password it was testing did not contain a number or uppercase
letter. Fix the spec to actually test the handling of both an invalid
and a valid password.