Add underline to hyperlinks in the web shell output
UBIDs in the web shell output are actually clickable links, which is a
very cool feature. However, it’s hard to notice at first glance, I only
realized it when I tried to copy one.

Let's add underlines to hyperlinks in the web shell output to make them
more visible.