Use write-through mode if the backing disk is write-through.
When the backing disk is already write-through, configuring ubiblk in
write-through mode adds minimal overhead, while having some performance
advantages: In this mode, the guest stops issuing flush requests because
the device reports no volatile cache. This can improve workloads such as
Postgres transactions.

On standard-8 VMs hosted on an AX-162 no-RAID and with a single IO
thread, pgbench with 32 clients ran significantly faster when ubiblk was
set to write-through. Statistics for reported TPS (based on 200 VMs per
mode):

| write_through |  p25  | median |  p75  |
|---------------|-------|--------|-------|
| true          | 12495 | 12858  | 13335 |
| false         |  9543 |  9908  | 10235 |

Note that this config will be effective on Ubiblk 0.2, and will be
ignored by Ubiblk 0.1.