Add support for renaming objects on the object's settings page
Support for renaming objects was previously added to the api/sdk/cli.
This adds support on the web. r.rename now works for both api and
web routes, and requires a new template_prefix option used to
construct the validation failure template path.

The template code for the new rename support is shared via a
new rename_object component.

In order to improve presentation and consistency, this adds a
"Danger Zone" heading to the settings pages previously lacking it.