Allow creating a PostgreSQL database without firewall rules in api/sdk/cli
The reason for allowing this is it allows you to create a locked-down
PostgreSQL database without parsing output of previous commands.
Previously, to create a locked-down database, you needed to create the
database, then remove the default firewall rules for it, then add your
own. Using this, you add the database without rules, and then add the
rules you need.