Reload services_lb before checking for CCM related changes
We regularly check to make sure the loadbalancer we have provisioned
in Clover is in sync with the LoadBalancers defined inside the k8s
cluster. In the current implementation the loadbalancer could get
cached and represent the clover lb wrongly and as a result, it would
increment the sync_kubernetes_service semaphores which would make no
changes.

With this commit, we reload the loadbalancer before checking for
changes making sure we have the latest change.