Fix issue in Kubernetes upgrade process
Prior to this commit, we would run the drain/delete command on the
last node of the cluster but in a setup where we have 1 control plane
node, by the time we are running the delete command, we have
already run the kubeadm reset resulting in errors since there are
no kubeconfigs there.

With this commit, we make sure we are choosing
a functional node and not a draining one.