Support configurable labels for links on admin site
Previously, links always were displayed as ubids. This changes it
to call the admin_label method on the object. By default, this will
call name on the object if it is defined, otherwise, it will use
the ubid.

Use a custom link label of email for Account, since that should be
more useful than the account name.

This makes many object types automatically have useful link labels.
Objects that don't implement name can override admin_label
appropriately. Just be careful not to introduce an N+1 query issue
when doing so.