Avoid unsynchronized setting of monitorable resource session to nil
Merge the process_event_loop method into the check_pulse method. Have the
pulse thread change a local variable if there was an error during the event
loop.  Have the monitor thread (not the pulse thread) clear the session,
after the pulse thread has ended, if the event loop failed.

Don't sleep until run_event_loop is true. With the merged code,
run_event_loop would be set immediately after the start of the thread, so
there is no reason to sleep. Even before this change, I don't think there
was a reason to sleep previously.