Check CSRF token on admin site
When the admin site was originally added, the only forms were
Rodauth forms, and Rodauth handled the CSRF checks.  However, as
we are adding other forms, we need to check the CSRF even for
non-Rodauth forms.

Use the same approach as in Clover, where Rodauth no longer checks
the CSRF, and the CSRF is checked before calling r.rodauth.