Add Pay Now button to unpaid invoices
Unfortunately, credit cards sometimes encounter issues with offline
recurring payments because they require 3D Secure. Several customers
have experienced this problem.

Currently, we manually create payment links for them after the invoice
is finalized at the start of the month. Once they complete the payment,
we mark the invoice as paid.

To automate this, I've added a "Pay Now" button to unpaid invoices.
Clicking it creates a checkout session and redirects the customer to
Stripe’s hosted payment page. This uses logic similar to adding credit
cards, but instead of "setup" mode, we now use "payment" mode. On the
success URL, we check the status of the session and the invoice ID to
mark the invoice as paid.