Use link instead of button for Vm/PostgresResource restart button on admin site
The link takes you to a separate page with the button. This is done to
avoid accidental clicks.

As we plan to support additional actions in the future that will want to use
the same framework, I decided to DRY things up. You now register supported
actions per class in an OBJECT_ACTIONS hash. The object page display links to
each action-specific page with the button to perform the action. This approach
will allow for model-specific actions with potentially only a single line of
code.