Skip runner pool if feature flag is enabled
We apply special allocation logic for certain GitHub installations, and
in some cases, we can provision an alien runner for them.

However, if the pool has available runners, it uses a VM from the pool
instead of provisioning a new one, which bypasses the special logic.

With this feature flag, we can skip the pool and always provision a new
VM for specific installations.