Remove role="form" attribute on form tags
It is redundant. role="form" should be set on non-form tags that
operate as forms, and if a form tag is used that doesn't operate
as a form, you could potentially use a different role, but the
assumption is that a form tag has a form role, so there is no
benefit to being explicit about it.