Attempt to fix nondeterminism in repartition stale spec
The main fix here is to the synchronization between the threads
that notify for partition numbers 2 and 3. Don't send the notify
for partition 2 until after the notification for partition 3 has
been received and applied by the repartitioner and checked by the
spec.

While here, use 3 separate queues for synchronization, instead
of only 2 queues, to rule out the possibility of queue unexpected
queue reuse causing the problem.

Add expectations for each queue pop and thread join, so that if
there is a timeout expiration, you can better diagnose the cause
of the spec failure.

This passed 100 spec runs without failure locally, where the
previous code generally didn't get through 30 spec runs without
failure locally.