Support creating private subnet when creating virtual machine
This adds a "New Private Subnet" option to the Private Subnet select
box. When chosen, the "Private Subnet Name" input displays. When
filled out correctly, it creates the virtual machine in the newly
created private subnet.

This uses a CSS approach instead of a Javascript approach for
handling the display of the "Private Subnet Name" input when the
"New Private Subnet" option is selected. This type of CSS is now
widely supported by browsers: https://caniuse.com/css-has