Fix repartitioner invalid partition spec nondeterminism
This attempts to fix the following error seen in CI:

```
  1) Repartitioner#listen emits and otherwise ignores invalid partition numbers
     Failure/Error: expect(mp.strand_id_range).to eq("00000000-0000-0000-0000-000000000000".."ffffffff-ffff-ffff-ffff-ffffffffffff")
       expected: "00000000-0000-0000-0000-000000000000".."ffffffff-ffff-ffff-ffff-ffffffffffff"
            got: "00000000-0000-0000-0000-000000000000"..."80000000-0000-0000-0000-000000000000"
```

I don't see how it is possible for this error to occur unless
something else is notifying on the same channel, and we are using
separate channels per spec.  However, since this doesn't affect
what we are actually testing for in this spec, avoid the
issue by not listening until after checking the initial strand
range. Also, set max_partition to 1 in the spec, so even if there
was an repartitioner notifying for partition number 2, it would
still be considered invalid and fail.