Detect the volume device filesystem using blkid
Prior to this commit, the logic to detect whether we need to create
filesystem for a device was flawed and it relied on the existence
of loop device or if it was copied across nodes or not.

Now with this commit, we will use blkid which is a reliable tool
to detect the filesystem and act according to it

Also some validation checks are added to make sure we are working
with the filesystems we support.