Show link to SSH public key management page on VM create page
If the user has not already registered an SSH public key, and they
have permission to do so, show a link to the page. This makes it
more likely users will be aware of this feature.

If the project has a registered SSH public key, we don't need to
explain the benefit of an SSH public key, so hide the description
in that case.