Allow creating a PostgreSQL database connected to an existing private subnet in the web UI
If there are existing private subnets created, this shows a select
box to select one. If selected, that private subnet is connected to
the private subnet created for the PostgreSQL database. This will
allow virtual machines in the connected subnet to connect to the
PostgreSQL database over the private IP address, allowing you to
better restrict access to the PostgreSQL database from the internet.