Support postgres private DNS in the web UI
If the postgres resource subnet is connected to another subnet,
support showing connection info with the private DNS name. It's
possible we should support it in all cases, even if not connected to
another subnet, since the user can add VMs directly to the postgres
resource subnet, but I think we want to encourage using connected
subnets instead of adding resources to postgres resource subnets.

Update the specs to correctly test the DNS names.