Allow connecting to postgres private subnets in the api/cli/sdk
For the cli, I choose to use a separate option for this. While it
currently wouldn't be ambiguous, as soon as we restrict private
subnet connection to private subnets in the same location, users
will be able to provide a private subnet name without a location,
and a postgres resource ubid is an allowable private subnet name.