Allow disconnecting from postgres private subnets in the api/cli/sdk
Similar to connecting, this uses a separate option. Rather than copying
the connect route code to disconnect and modify it, since the majority
of the route code is the same, this combines the private subnet connect
and disconnect routes using a private_subnet_connection_action helper
method.

In order to tell the difference between postgres and private subnet
ids, we need to keep in ubid format so we can check the first two
characters. Add ubid typecast param and symbol matcher to support that.