Use spot instances for alien runners
On-demand AWS instances are about 3x more expensive than our own virtual
machines. To make this feature viable, we plan to use spot instances for
alien runners. The average spot cost for m7a is still close to our
runner prices, but since we’ll only use them for spillover spiky
workloads, that should be fine.

If we start spilling too many runners, that’s a clear signal that we
need to add more hosts to our fleet.