Attempt to fix dispatcher stale spec nondeterminism
Don't push to the queue until after the prepared statement is setup,
instead of before.

Don't set a low stale_seconds, keep the default and manually
backdate the times for partitions 2 and 3 to show them as in the
past.

Drop the separate thread to manually notify for the current
partition, it isn't necessary.

Rename some variables and add comments to better describe the
synchronization and purpose of the tests.