Make authorized_object and related methods require location_id
Always use the location_id as a filter. This prevents the
accidental omission of location_id, which would allow objects
to be related to objects in other locations.

There is one current case where no location is provided, and that
is when creating a load balancer, where the load balancer doesn't
have a location by itself, its location is derived from the
related private subnet. Handle this by providing Sequel::NOTNULL
as a filter, as a way to be explicit that any location is allowed.
As location_id is a NOT NULL column, the PostgreSQL optimizer will
remove the location_id IS NOT NULL predicate check.

This moves the location_id keyword argument to be before the optional
keyword arguments in the methods, now that it is required.