Remove `security: - BearerAuth: []` from a couple openapi.yml endpoints
I'm not sure why these are in only these two endpoints. This is already
a top-level openapi.yml setting that sets a default, so you would only
want to specify it for specific endpoints to override the default.
However, we aren't doing that, since the endpoint values match the
top-level setting.