Add a prog to remove a ubiblk backend.
It waits for volumes using the backend to destroy, and then removes the
ubiblk installation by sending a command to rhizome and updating the
database.

We'll receive a page if the volumes using the backend don't go away in 4
hours. This seems to be a reasonable time to wait for github runner VMs
to finish their jobs.

At the moment, for general purpose VMs it is expected that an operator
manually upgrades VMs to a newer ubiblk backend before removing older
installations. We'll consider automating this in future.