Set nil when billing info attributes are empty
tax_id, company_name, postal_code, state, and city are not required on
the billing info update form. When they are not provided, we should set
them to nil instead of an empty string.

Since an empty string is truthy in Ruby, it causes issues when we try to
check whether billing info details are present.