Add e2e test to check the data hash after data migration
Number of worker nodes in the e2e test cluster is increased to two
so we can test the data migration. During the test, first we write
some random data to the pod's mount location, then cordon its node
in order to trigger the migration and then wait for the new pod
to be running and then finally check the hash after migration.