Add UI for K8s Nodepool Scaling
This commit adds a select component on the Setting page of
Kubernetes clusters that would allow changing the size of the first
(and only) node pool of the cluster.

In the future, when we implement the nodepool management, we can
have the resize logic in the Nodepools tab, alongside the rename.